+++ SHPViewerStudy/App/UI/CPanelRight.cpp
@@ -8,10 +8,6 @@ END_MESSAGE_MAP()
 
 bool CRightPanel::Create(CWnd* pParent, UINT nID, CRect& rect)
 {
-    //m_bgBrush.CreateSolidBrush(RGB(240, 240, 240));
-    // WS_VISIBLE æ¯¿Ã Ω√¿€ = ¥›»˘ ªÛ≈¬
-    //m_clientWidth  = rect.Width();
-    //m_clientHeight = rect.Height();
     return CWnd::Create(AfxRegisterWndClass(CS_HREDRAW | CS_VREDRAW, ::LoadCursor(nullptr, IDC_ARROW), (HBRUSH)(COLOR_BTNFACE + 1)), _T(""), WS_CHILD | WS_CLIPCHILDREN, CRect(0, 0, 10, 10), pParent, nID) == TRUE;
 }
 
@@ -24,18 +20,19 @@ int CRightPanel::OnCreate(LPCREATESTRUCT lp)
 
 void CRightPanel::Resize()
 {
-    //m_clientWidth  = screenWidth;
-    //m_clientHeight = screenHeight;
     int32_t panelWidth = GetWidth();
     int32_t marginX = static_cast<int32_t>(panelWidth * 0.05f);
 
     MoveWindow(m_uiSize.clientWidth - panelWidth, 0, panelWidth, m_uiSize.clientHeight);
     m_staticInfo.MoveWindow(marginX, marginX, panelWidth - marginX * 2, m_uiSize.clientHeight - marginX * 2);
     
-    int32_t fontSize = std::max(10, m_uiSize.clientHeight / 40);
-    m_font.DeleteObject();
-    m_font.CreateFont(-fontSize, 0, 0, 0, FW_NORMAL, FALSE, FALSE, FALSE, DEFAULT_CHARSET, OUT_DEFAULT_PRECIS, CLIP_DEFAULT_PRECIS, CLEARTYPE_QUALITY, DEFAULT_PITCH | FF_SWISS, _T("Segoe UI"));
-    m_staticInfo.SetFont(&m_font);
+    
+    if (m_uiSize.isFontChanged) {
+        int32_t fontSize = std::max(10, m_uiSize.clientHeight / 40);
+        m_font.DeleteObject();
+        m_font.CreateFont(-fontSize, 0, 0, 0, FW_NORMAL, FALSE, FALSE, FALSE, DEFAULT_CHARSET, OUT_DEFAULT_PRECIS, CLIP_DEFAULT_PRECIS, CLEARTYPE_QUALITY, DEFAULT_PITCH | FF_SWISS, _T("Segoe UI"));
+        m_staticInfo.SetFont(&m_font);
+    }
 }
 
 void CRightPanel::SetPickingInfo(const CString& text)
